Migrate common PID to new logic

+++ src/main/flight/pid.h
@@ -142,7 +142,7 @@ void updatePIDCoefficients(void);
 void pidController(void);
 
 float pidRateToRcCommand(float rateDPS, uint8_t rate);
-int16_t pidAngleToRcCommand(float angleDeciDegrees, int16_t maxInclination);
+float pidAngleToRcCommand(float angleDeciDegrees, float maxInclination);
 
 enum {
     HEADING_HOLD_DISABLED = 0,
